+++ frontend/src/components/ProblemForm/index.tsx
@@ -1,48 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useMutation } from 'react-query';
 import { problemAPI } from '../../api/problem';
 import InputFile from './InputFile';
-
-/* TODO : 이거 분리하기, Platform도 Object로 뽑아놓기 */
-const levelList = {
-  boj: ['골드', '실버', '브론즈2 이상'],
-  programmers: ['Level3', 'Level2', 'Level1'],
-  hackerrank: ['hard', 'medium', 'easy'],
-  default: ['== =='],
-};
-
-interface LevelProps {
-  platform: string;
-}
-
-/* TODO : 컴포넌트 다른파일로 분리하기 */
-const LevelSelect = ({ platform }: LevelProps) => {
-  const [options, setOptions] = useState<string[]>(levelList.default);
-
-  useEffect(() => {
-    switch (platform) {
-      case 'boj':
-        setOptions(levelList.boj);
-        break;
-      case 'programmers':
-        setOptions(levelList.programmers);
-        break;
-      case 'hackerrank':
-        setOptions(levelList.hackerrank);
-        break;
-      default:
-        break;
-    }
-  }, [platform]);
-
-  return (
-    <select className="translate-all h-11 scale-95 rounded-xl border-[1px] p-3 shadow-sm" id="level" name="level">
-      {options.map((option) => (
-        <option value={option}>{option}</option>
-      ))}
-    </select>
-  );
-};
+import LevelSelect from './LevelSelect';
 
 interface ProblemFormProps {
   close: () => void;
@@ -109,7 +69,7 @@ const ProblemForm = ({ close }: ProblemFormProps) => {
           name="link"
         />
 
-        <InputFile updateUrl={handleImgUrl} />
+        <InputFile updateImg={updateImg} />
 
         <div className="m-2 flex justify-around gap-4">
           <button
